基本功能实现

+++ todolist.py
@@ -1,6 +1,6 @@
 import os
 
-from flask import Flask, request, render_template, redirect, session, url_for
+from flask import Flask, request, render_template, redirect, session, url_for, flash
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 from flask_bootstrap import Bootstrap
@@ -54,45 +54,39 @@ class UserForm(FlaskForm):
 
 
 @app.route('/')
-def hello_world():
-    # if session.get('username') and session.get('password'):
-    #     user = User.query.get(session.get('username'))
-    #     if user:
-    #         tasks = Task.query.filter_by(username=user.username).all()
-    #         return render_template('extend.html',tasks=tasks,user=user.username)
+def index():
+    if session['user'] is not None:
+        user = User.query.get(session['user'])
+        tasks = user.tasks
+        return render_template('index.html', username=user.username, tasks=tasks)
     return render_template('index.html', form=UserForm())
 
-@app.route('/user',methods=['POST'])
+
+@app.route('/user', methods=['POST'])
 def user_admin():
-
-
-@app.route('/signup', methods=['POST'])
-def signup():
-    username = request.form['username']
-    password = request.form['password']
-    if not User.query.get('username'):
-        user = User(username=username, password=password)
-        db.session.add(user)
-        db.session.commit()
-        session['username'] = username
-        session['password'] = password
-        return redirect(url_for('hello_world'))
-    return "用户已存在"
-
-
-@app.route('/login', methods=['POST'])
-def login():
-    username = request.form['username']
-    password = request.form['password']
-    user = User.query.get(username)
-    if user:
-        if user.password != password:
-            return '密码错误！'
+    form = UserForm()
+    if form.validate_on_submit():
+        login = form.login.data
+        user = User.query.filter_by(username=form.username.data).first()
+        msg = None
+        if login:
+            if user is None:
+                msg = '用户不存在！'
+            elif user.password != form.password.data:
+                msg = '用户名或密码错误！'
+            else:
+                session['user'] = user.username
         else:
-            session['username'] = request.form['username']
-            session['password'] = request.form['password']
-            return redirect(url_for('hello_world'))
-    return "用户不存在！"
+            if user is None:
+                msg = '注册成功！'
+                user = User(username=form.username.data, password=form.password.data)
+                db.session.add(user)
+                db.session.commit()
+                session['user'] = user.username
+            else:
+                msg = '用户名已存在'
+        if msg is not None: flash(msg)
+    return redirect(url_for('index'))
 
 
 @app.route('/logout')
@@ -117,26 +111,20 @@ def add_task():
 def done(id):
     # id = request.form['id']
     task = Task.query.get(id)
-    if task:
+    # msg = None
+    if task and session['user'] and session['user'] == task.username:
         task.finished = True
         db.session.commit()
-        return redirect(url_for('hello_world'))
-    return "error"
+    return redirect(url_for('index'))
 
 
 @app.route('/del/<id>')
 def delete(id):
     task = Task.query.get(id)
-    if task:
+    if task and session['user'] and session['user'] == task.username:
         db.session.delete(task)
         db.session.commit()
-        return redirect(url_for('hello_world'))
-    return "error"
-
-
-@app.route('/about')
-def about():
-    return render_template('about.html')
+    return redirect(url_for('index'))
 
 
 if __name__ == '__main__':
